Renamed StylableHBox to HBox. Added dir prop. Removed convoluted for-loop
