+++ packages/wix-ui-core/stories/HBox/index.ts
@@ -2,15 +2,16 @@ import * as React from 'react';
 import createStory from '../create-story';
 
 import {HBox} from '../../src/components/StylableHBox';
-import * as HBoxSource from '!raw-loader!../../src/components/StylableHBox/HBox.tsx';
+import * as HBoxSource from '!raw-loader!../../src/components/HBox/HBox.tsx';
 import style from './style.st.css';
 
 const children = new Array(5).fill(undefined).map(() => React.createElement('div', {}, 'hello'));
+children.push(React.createElement('div', {}, 'Hello'));
 
 export const story = () => createStory({
   category: 'Components',
   name: 'HBox',
-  storyName: 'StylableHBox',
+  storyName: 'HBox',
   component: HBox,
   componentProps: (setState) => ({
     ...style('root'),
